+++ packages/runner-ct/src/app/KeyboardHelper.scss
@@ -1,0 +1,42 @@
+.keyboard-helper-container {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 100%;
+
+  .keyboard-helper { 
+    padding: 8px 32px;
+    color: #555;
+    // background: white;
+    font-family: "Muli", "Helvetica Neue", "Arial", sans-serif;
+    font-size: 13px;
+    width: 224px;
+
+    .keyboard-shortcut {
+      display: flex;
+      margin-top: 8px;
+      height: 23px;
+      justify-content: space-between;
+
+      .shortcut { 
+        display: flex;
+
+        .key {
+          display: flex;
+          font-family: sans-serif; // display keys symbols correctly
+          justify-content: center;
+          align-items: center;
+          border: 1px solid rgba(255, 255, 255, 0.4);
+          height: 23px;
+          min-width: 23px;
+          margin-right: 4px;
+          padding: 0px 4px;
+          font-size: 0.8125rem;
+          border-radius: 4px;
+          pointer-events: none;
+          background-color: #ddd;
+        }
+      }
+    }
+  }
+}
